// src/types/positionbook.ts
export type PositionBook = {
  //   stat: string;
  uid: string;
  actid: string;
  exch: string;
  tsym: string;
  s_prdt_ali: string;
  prd: string;
  token: string;
  netqty: string;
  netavgprc: string;
  dname: string;
  cname: string | undefined;
  // daybuyqty: string;
  // daysellqty: string;
  // daybuyavgprc: string;
  // daysellavgprc: string;
  // daybuyamt: string;
  // daysellamt: string;
  // cfbuyqty: string;
  // cforgavgprc: string;
  // cfsellqty: string;
  // cfbuyavgprc: string;
  // cfsellavgprc: string;
  // cfbuyamt: string;
  // cfsellamt: string;
  // openbuyqty: string;
  // opensellqty: string;
  // openbuyamt: string;
  // opensellamt: string;
  // openbuyavgprc: string;
  // opensellavgprc: string;
  mult: string;
  lp: string;
  rpnl: string;
  urmtom: string;
  bep: string;
  pp: string;
  prcftr: string;
  ti: string;
  ls: string;
};